// frontend/src/client/types.gen.ts
export type OptionsGalleryResult = {
    id: string;
    course_id: string;
    location?: string | null;
    duration: string;
    googleMapsUrl?: string | null;
    startDate?: string | null;
    Qualification?: string | null;
};